+++ packages/gatsby-plugin-breadcrumbs/src/usePageBreadcrumbsContext.js
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+
+import pageBreadcrumbsContext from "./pageBreadcrumbsContext";
+
+export default function usePageBreadcrumbsContext() {
+  return useContext(pageBreadcrumbsContext);
+}