+++ packages/gatsby-plugin-cookie-consent/src/CookieConsentProvider.js
@@ -11,7 +11,12 @@ CookieConsentProvider.propTypes = {
 export default function CookieConsentProvider({ children }) {
   const { whitelist } = useCookieConsentSettings();
 
-  useBlockScripts({ whitelist });
+  if (whitelist) {
+    console.warn(
+      `Setting a whitelist through useCookieConsentSettings is deprecated and <CookieConsentProvider> is no longer used. Please remove it from your app and use plugin options to set the whitelist.`,
+    );
+    useBlockScripts({ whitelist });
+  }
 
   return <>{children}</>;
 }