+++ packages/gatsby-plugin-i18next/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.0](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.8...v1.2.0) (2022-09-25)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.8](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.7...v1.2.0-next.8) (2022-09-12)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.7](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.6...v1.2.0-next.7) (2022-08-24)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.6](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.5...v1.2.0-next.6) (2022-08-24)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.5](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.4...v1.2.0-next.5) (2022-06-03)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.4](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.3...v1.2.0-next.4) (2022-06-03)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.3](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.2...v1.2.0-next.3) (2022-06-03)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.2](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.1...v1.2.0-next.2) (2022-06-02)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.1](https://github.com/whitespace-se/gatsby-packages/compare/v1.2.0-next.0...v1.2.0-next.1) (2022-05-18)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
+## [1.2.0-next.0](https://github.com/whitespace-se/gatsby-packages/compare/v1.1.0...v1.2.0-next.0) (2022-05-18)
+
+**Note:** Version bump only for package @whitespace/gatsby-plugin-i18next
+
+
+
+
+
 ## [1.1.0](https://github.com/whitespace-se/gatsby-packages/compare/v1.1.0-next.3...v1.1.0) (2022-05-09)
 
 **Note:** Version bump only for package @whitespace/gatsby-plugin-i18next